Break up setup_nodes into setup_services, for individial services, and setup_nodes, for initializing different simulator configurations
